# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "El_Llibres API",
        "description": "API for managing Books",
        "contact": {
            "email": "pol.romeu@cirvianum.cat"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/likes": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Retrieve a list of likes",
                "description": "List all Likes.",
                "operationId": "297b9c3ea341e74585a2ae193ef6dcee",
                "responses": {
                    "200": {
                        "description": "A list of likes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "likes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "book_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Create a new like",
                "description": "Create a new like.",
                "operationId": "511c8161eb44f7b3479a914692d975ee",
                "requestBody": {
                    "description": "Like details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Like created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/likes/{id}": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Retrieve a specific like",
                "description": "Retrieve a specific like by ID.",
                "operationId": "cc6d000adc2ca5c218da958d798e920a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Like ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "book_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Like not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Likes"
                ],
                "summary": "Update a specific like",
                "description": "Update a specific like.",
                "operationId": "4fca184960c72abcfe28c03f60e7e12b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Like ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated like details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Like updated successfully"
                    },
                    "404": {
                        "description": "Like not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Likes"
                ],
                "summary": "Delete a specific like",
                "description": "Delete a specific like.",
                "operationId": "e02fff678edcb4bc50ab9c57d7e4531d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Like ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like deleted successfully"
                    },
                    "404": {
                        "description": "Like not found"
                    }
                }
            }
        },
...
    },
...
}